Move tenant resources to dedicated pkg
