+++ internal/provider/provider.go
@@ -15,6 +15,7 @@ import (
 	"github.com/auth0/terraform-provider-auth0/internal/auth0/connection"
 	"github.com/auth0/terraform-provider-auth0/internal/auth0/organization"
 	"github.com/auth0/terraform-provider-auth0/internal/auth0/resourceserver"
+	"github.com/auth0/terraform-provider-auth0/internal/auth0/tenant"
 )
 
 var version = "dev"
@@ -96,7 +97,7 @@ func New() *schema.Provider {
 			"auth0_email":                      newEmail(),
 			"auth0_email_template":             newEmailTemplate(),
 			"auth0_user":                       newUser(),
-			"auth0_tenant":                     newTenant(),
+			"auth0_tenant":                     tenant.NewResource(),
 			"auth0_role":                       newRole(),
 			"auth0_log_stream":                 newLogStream(),
 			"auth0_branding":                   newBranding(),
@@ -115,7 +116,7 @@ func New() *schema.Provider {
 			"auth0_connection":      connection.NewDataSource(),
 			"auth0_organization":    organization.NewDataSource(),
 			"auth0_resource_server": resourceserver.NewDataSource(),
-			"auth0_tenant":          newDataTenant(),
+			"auth0_tenant":          tenant.NewDataSource(),
 		},
 	}
